Fix compiler errors in Indexer due to version update
* cstring header automatically included in Visual Studio but not in g++
* filesystem library moved from experimental namespace to std namespace
* Replace non-standard itoa function with to_string from C++11
* Change directory separator to forward-slash

+++ Indexer/hash.cpp
@@ -23,7 +23,7 @@ unsigned long long int power(unsigned long long int x, unsigned long long int y,
 namespace indexer 
 {
 
-	size_t hash(std::string str)
+	unsigned int hash(std::string str)
 	{
 		size_t hash_value = 0, prime = 1000000007;
